+++ realm/realm-library/src/main/java/io/realm/log/LogLevel.java
@@ -27,6 +27,12 @@ package io.realm.log;
 public class LogLevel {
 
     /**
+     * Not implemented. It forces javac to generate the JNI header file for this class.
+     */
+    @SuppressWarnings("unused")
+    private native void forceJNIHeaderGeneration();
+
+    /**
      * The ALL has the lowest possible rank and is intended to turn on all logging.
      */
     public static final int ALL = 1;